+++ x/btcstkconsumer/keeper/consumer_registry.go
@@ -51,13 +51,13 @@ func (k Keeper) GetConsumerRegister(ctx context.Context, consumerID string) (*ty
 	return &consumerRegister, nil
 }
 
-// GetConsumerRegistryMaxMultiStakedFps gets the max_multi_staked_fps from the consumer registry for a given consumer ID
+// GetConsumerRegistryMaxMultiStakedFps gets the consumer_max_multi_staked_fps from the consumer registry for a given consumer ID
 func (k Keeper) GetConsumerRegistryMaxMultiStakedFps(ctx context.Context, consumerID string) (uint32, error) {
 	consumerRegister, err := k.GetConsumerRegister(ctx, consumerID)
 	if err != nil {
 		return 0, err
 	}
-	return consumerRegister.MaxMultiStakedFps, nil
+	return consumerRegister.ConsumerMaxMultiStakedFps, nil
 }
 
 // GetAllRegisteredConsumerIDs gets all consumer IDs that registered to Babylon